Enum getopts::Fail

pub enum Fail {
    ArgumentMissing(String),
    UnrecognizedOption(String),
    OptionMissing(String),
    OptionDuplicated(String),
    UnexpectedArgument(String),
}

The type returned when the command line does not conform to the expected format. Use the Show implementation to output detailed information.

Variants

ArgumentMissing

The option requires an argument but none was passed.

UnrecognizedOption

The passed option is not declared among the possible options.

OptionMissing

A required option is not present.

OptionDuplicated

A single occurrence option is being used multiple times.

UnexpectedArgument

There's an argument being passed to a non-argument option.

Methods

impl Fail

fn to_err_msg(self) -> String

Convert a Fail enum into an error string.