Function getopts::each_split_within

pub(crate) fn each_split_within(desc: &str, lim: usize) -> Vec<String>

Splits a string into substrings with possibly internal whitespace, each of them at most lim bytes long, if possible. The substrings have leading and trailing whitespace removed, and are only cut at whitespace boundaries.