Function getopts::getopts

pub fn getopts(args: &[String], optgrps: &[OptGroup]) -> Result

Parse command line arguments according to the provided options.

On success returns Ok(Matches). Use methods such as opt_present opt_str, etc. to interrogate results.

Panics

Returns Err(Fail) on failure: use the Show implementation of Fail to display information about it.