Function getopts::reqopt

pub fn reqopt(short_name: &str, long_name: &str, desc: &str, hint: &str) -> OptGroup

Unstable
: use the crates.io getopts library instead

Create a long option that is required and takes an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option