Function getopts::each_split_within

fn each_split_within<'a, F>(ss: &'a str, lim: usize, it: F) -> bool where F: FnMut(&'a str) -> bool

Splits a string into substrings with possibly internal whitespace, each of them at most lim bytes long. The substrings have leading and trailing whitespace removed, and are only cut at whitespace boundaries.

Note: Function was moved here from std::str because this module is the only place that uses it, and because it was too specific for a general string function.

Panics

Panics during iteration if the string contains a non-whitespace sequence longer than the limit.