Struct getopts::Matches

pub struct Matches {
    pub free: Vec<String>,
    // some fields omitted
}

The result of checking command line arguments. Contains a vector of matches and a vector of free strings.

Fields
free: Vec<String>

Free string fragments

Methods

impl Matches

pub fn opt_defined(&self, nm: &str) -> bool

Returns true if an option was defined

pub fn opt_present(&self, nm: &str) -> bool

Returns true if an option was matched.

pub fn opt_count(&self, nm: &str) -> usize

Returns the number of times an option was matched.

pub fn opt_positions(&self, nm: &str) -> Vec<usize>

Returns a vector of all the positions in which an option was matched.

pub fn opts_present(&self, names: &[String]) -> bool

Returns true if any of several options were matched.

pub fn opts_str(&self, names: &[String]) -> Option<String>

Returns the string argument supplied to one of several matching options or None.

pub fn opt_strs(&self, nm: &str) -> Vec<String>

Returns a vector of the arguments provided to all matches of the given option.

Used when an option accepts multiple values.

pub fn opt_strs_pos(&self, nm: &str) -> Vec<(usize, String)>

Returns a vector of the arguments provided to all matches of the given option, together with their positions.

Used when an option accepts multiple values.

pub fn opt_str(&self, nm: &str) -> Option<String>

Returns the string argument supplied to a matching option or None.

pub fn opt_default(&self, nm: &str, def: &str) -> Option<String>

Returns the matching string, a default, or None.

Returns None if the option was not present, def if the option was present but no argument was provided, and the argument if the option was present and an argument was provided.

pub fn opt_get<T>(&self, nm: &str) -> Result<Option<T>, T::Err> where
T: FromStr,

Returns some matching value or None.

Similar to opt_str, also converts matching argument using FromStr.

pub fn opt_get_default<T>(&self, nm: &str, def: T) -> Result<T, T::Err> where
T: FromStr,

Returns a matching value or default.

Similar to opt_default, except the two differences. Instead of returning None when argument was not present, return def. Instead of returning &str return type T, parsed using str::parse().