Struct getopts::Options

pub struct Options { /* private fields */ }

A description of the options that a program can handle.

Implementations

impl Options

pub fn new() -> Options

Create a blank set of options.

pub fn parsing_style(&mut self, style: ParsingStyle) -> &mut Options

Set the parsing style.

pub fn long_only(&mut self, long_only: bool) -> &mut Options

Set or clear “long options only” mode.

In “long options only” mode, short options cannot be clustered together, and long options can be given with either a single “-” or the customary “–”. This mode also changes the meaning of “-a=b”; in the ordinary mode this will parse a short option “-a” with argument “=b”; whereas in long-options-only mode the argument will be simply “b”.

pub fn opt(
&mut self,
short_name: &str,
long_name: &str,
desc: &str,
hint: &str,
hasarg: HasArg,
occur: Occur
) -> &mut Options

Create a generic option group, stating all parameters explicitly.

pub fn optflag(
&mut self,
short_name: &str,
long_name: &str,
desc: &str
) -> &mut Options

Create a long option that is optional and does not take an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

pub fn optflagmulti(
&mut self,
short_name: &str,
long_name: &str,
desc: &str
) -> &mut Options

Create a long option that can occur more than once and does not take an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

pub fn optflagopt(
&mut self,
short_name: &str,
long_name: &str,
desc: &str,
hint: &str
) -> &mut Options

Create a long option that is optional and takes an optional argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

pub fn optmulti(
&mut self,
short_name: &str,
long_name: &str,
desc: &str,
hint: &str
) -> &mut Options

Create a long option that is optional, takes an argument, and may occur multiple times.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

pub fn optopt(
&mut self,
short_name: &str,
long_name: &str,
desc: &str,
hint: &str
) -> &mut Options

Create a long option that is optional and takes an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

pub fn reqopt(
&mut self,
short_name: &str,
long_name: &str,
desc: &str,
hint: &str
) -> &mut Options

Create a long option that is required and takes an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

pub fn parse<C: IntoIterator>(&self, args: C) -> Result where
C::Item: AsRef<OsStr>,

Parse command line arguments according to the provided options.

On success returns Ok(Matches). Use methods such as opt_present opt_str, etc. to interrogate results.

Panics

Returns Err(Fail) on failure: use the Debug implementation of Fail to display information about it.

pub fn short_usage(&self, program_name: &str) -> String

Derive a short one-line usage summary from a set of long options.

pub fn usage(&self, brief: &str) -> String

Derive a formatted message from a set of options.

pub fn usage_with_format<F: FnMut(&mut dyn Iterator<Item = String>) -> String>(
&self,
formatter: F
) -> String

Derive a custom formatted message from a set of options. The formatted options provided to a closure as an iterator.