Struct getopts::Options

pub struct Options {
    // some fields omitted
}

A description of the options that a program can handle

Methods

impl Options

fn new() -> Options

Create a blank set of options

fn opt(&mut self, short_name: &str, long_name: &str, desc: &str, hint: &str, hasarg: HasArg, occur: Occur) -> &mut Options

Create a generic option group, stating all parameters explicitly

fn optflag(&mut self, short_name: &str, long_name: &str, desc: &str) -> &mut Options

Create a long option that is optional and does not take an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

fn optflagmulti(&mut self, short_name: &str, long_name: &str, desc: &str) -> &mut Options

Create a long option that can occur more than once and does not take an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

fn optflagopt(&mut self, short_name: &str, long_name: &str, desc: &str, hint: &str) -> &mut Options

Create a long option that is optional and takes an optional argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

fn optmulti(&mut self, short_name: &str, long_name: &str, desc: &str, hint: &str) -> &mut Options

Create a long option that is optional, takes an argument, and may occur multiple times.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

fn optopt(&mut self, short_name: &str, long_name: &str, desc: &str, hint: &str) -> &mut Options

Create a long option that is optional and takes an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

fn reqopt(&mut self, short_name: &str, long_name: &str, desc: &str, hint: &str) -> &mut Options

Create a long option that is required and takes an argument.

short_name - e.g. "h" for a -h option, or "" for none

long_name - e.g. "help" for a --help option, or "" for none

desc - Description for usage help

hint - Hint that is used in place of the argument in the usage help, e.g. "FILE" for a -o FILE option

fn parse(&self, args: &[String]) -> Result

Parse command line arguments according to the provided options.

On success returns Ok(Matches). Use methods such as opt_present opt_str, etc. to interrogate results.

Panics

Returns Err(Fail) on failure: use the Show implementation of Fail to display information about it.

fn short_usage(&self, program_name: &str) -> String

Derive a short one-line usage summary from a set of long options.

fn usage(&self, brief: &str) -> String

Derive a usage message from a set of long options.